adjustments

+++ src/modules/courses/services/DeleteCourseService.ts
@@ -1,3 +1,4 @@
+import Registration from '@modules/registration/models/registration';
 import AppError from '@shared/errors/AppError';
 import Course from '../models/course';
 
@@ -7,6 +8,14 @@ class DeleteCourseService {
 
     if (!course) {
       throw new AppError('Course not found.');
+    }
+
+    const registrations = await Registration.find({ course_id: course.id });
+
+    if (registrations.length != 0) {
+      throw new AppError(
+        'A course cannot be deleted with students enrolled in it.',
+      );
     }
 
     await course.delete();
